Changed some core mechanics of the game, pictures get recolered now and added new kind of db (relational)
